Practica 15 Finished & Try Files for the Invoice Practice
The Files from the Invoice Practice are deleted:
- list.bin, list.txt, persons.dat, text.txt and test2.txt.

From Practica 15:
- Book.java:
On the getEditorial() function the function returns title instead for the editorial.

- MainWindow.form:
The XML file includes all the buttons are set, the design is finished but have some bugs (Needs to be fixed).

- MainWindow.java:
The ArrayList, DefaultTableModel and Book Libraries are imported.
The comments for the class are completed.
All the variables are initialized in the initComponents function.
The (function) completeInput() checks if the all Inputs have a value or not.
The (function) addBookButtonActionPerformed() adds a book inside the table.
The (function) removeBookButtonActionPerformed() remove a book inside the table.
The (function) getSelectedAge() returns the age at which the book is redirected.
The (function) getSelectedGenres() returns an array list with the selected genres.
All the general variables are declared at the final.